Adds vars_file include to building playbook
The playbook requires the existence of these vars, so
it makes sense to specify the include in the playbook itself.
Specifying it only in the playbook doesn't permit custom
error-handling messages (at least not until Ansible 2,
which will support exception handling via block/except),
so the Vagrantfile also includes an existence check.