Adds missing 'digest' dependency to Rakefile
It doesn't need to be in the Gemfile, but should be
included for running the `rake` task, since it calls
`Digest::SHA256`.

Whether it must be included may be dependent on Ruby versions,
and I'm not going to go down that rabbit hole now, but did
confirm that including it works in:

  * 2.1.2p95 (Ubuntu)
  * 2.1.5p273 (Debian)

and that's good enough for me. The long-term solution is to
remove the Ruby dependency all together.